add invIndex class, cause gson
